requirements.txt
Flask==3.0.0
requests==2.31.0
python-dotenv==1.0.0
googlemaps==4.10.0
pandas==2.1.3
openpyxl==3.1.2
beautifulsoup4==4.12.2
lxml==4.9.3
gspread==5.12.0
google-auth==2.23.4
supabase>=2.0.0,<3

